Add tests for quarkus.elasticsearch.health.enabled
