+++ integration-tests/elasticsearch-rest-client/src/test/java/io/quarkus/it/elasticsearch/profile/HealthDisabledTestProfile.java
@@ -1,0 +1,14 @@
+package io.quarkus.it.elasticsearch.profile;
+
+import java.util.Collections;
+import java.util.Map;
+
+import io.quarkus.test.junit.QuarkusTestProfile;
+
+public class HealthDisabledTestProfile implements QuarkusTestProfile {
+
+    @Override
+    public Map<String, String> getConfigOverrides() {
+        return Collections.singletonMap("quarkus.elasticsearch.health.enabled", "false");
+    }
+}
